Revert "исправление ошибок"
