+++ delivery-service/src/main/java/com/example/Main.java
@@ -1,0 +1,6 @@
+package com.example;
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
