+++ delivery-service/src/main/java/com/example/StatusCurrier.java
@@ -1,0 +1,7 @@
+package com.example;
+
+public enum StatusCurrier {
+    ORDERISTAKE,
+    ORDERONTHEWAY,
+    ORDERTRANSFERBUYER
+}
