+++ delivery-service/src/main/java/com/example/controllers/DeliveryController.java
@@ -1,0 +1,16 @@
+package com.example.controllers;
+
+public class DeliveryController {
+    public void createDeliver() {
+
+    }
+    public void changeDeliver() {
+
+    }
+    public void getDeliver() {
+
+    }
+    public void deleteDeliver() {
+
+    }
+}
