+++ delivery-service/src/main/java/com/example/models/Deliver.java
@@ -1,0 +1,12 @@
+package com.example.models;
+
+public class Deliver {
+    int id;
+   String name;
+   String surname;
+   int age;
+   int phoneNumber;
+
+   String statusOrder;
+
+}
