--- delivery-service/src/main/java/ru/liga/DeliveryServiceApplication.java
@@ -1,12 +1,0 @@
-package ru.liga;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DeliveryServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(DeliveryServiceApplication.class, args);
-    }
-
-}
