--- delivery-service/src/main/java/ru/liga/StatusCurrier.java
@@ -1,7 +1,0 @@
-package ru.liga;
-
-public enum StatusCurrier {
-    ORDERISTAKE,
-    ORDERONTHEWAY,
-    ORDERTRANSFERBUYER
-}
