--- delivery-service/src/main/java/ru/liga/controllers/DeliveryController.java
@@ -1,16 +1,0 @@
-package ru.liga.controllers;
-
-public class DeliveryController {
-    public void createDeliver() {
-
-    }
-    public void changeDeliver() {
-
-    }
-    public void getDeliver() {
-
-    }
-    public void deleteDeliver() {
-
-    }
-}
