--- delivery-service/src/main/java/ru/liga/models/Deliver.java
@@ -1,12 +1,0 @@
-package ru.liga.models;
-
-public class Deliver {
-    int id;
-   String name;
-   String surname;
-   int age;
-   int phoneNumber;
-
-   String statusOrder;
-
-}
