--- delivery-service/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-server.port=8081
