+++ dependency-bom/pom.xml
@@ -4,21 +4,20 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <groupId>ru.liga</groupId>
+        <groupId>com.example</groupId>
         <artifactId>ru-liga-food-service</artifactId>
-        <version>0.0.1</version>
+        <version>0.0.1-SNAPSHOT</version>
     </parent>
 
     <artifactId>dependency-bom</artifactId>
     <packaging>pom</packaging>
 
     <properties>
-        <spring.version>2.7.15</spring.version>
+        <spring.version>2.7.15-SNAPSHOT</spring.version>
         <java.version>11</java.version>
         <maven.compiler.soure>11</maven.compiler.soure>
         <maven.compiler.target>11</maven.compiler.target>
         <maven.compiler.plugin>3.11.0</maven.compiler.plugin>
-        <snakeyaml.version>2.2</snakeyaml.version>
     </properties>
 
 
@@ -38,15 +37,7 @@
             <dependency>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-starter-actuator</artifactId>
-                <version>${spring.version}</version>
             </dependency>
-            <!-- https://mvnrepository.com/artifact/org.yaml/snakeyaml -->
-            <dependency>
-                <groupId>org.yaml</groupId>
-                <artifactId>snakeyaml</artifactId>
-                <version>${snakeyaml.version}</version>
-            </dependency>
-
             <!--		<dependency>-->
             <!--			<groupId>org.springframework.security</groupId>-->
             <!--			<artifactId>spring-security-test</artifactId>-->
@@ -61,15 +52,11 @@
                     <dependency>
                         <groupId>org.springframework.boot</groupId>
                         <artifactId>spring-boot-starter-amqp</artifactId>
-                        <version>${spring.version}</version>
                     </dependency>
-
 
             <!--		<dependency>-->
             <!--			<groupId>org.springframework.boot</groupId>-->
             <!--			<artifactId>spring-boot-starter-data-jpa</artifactId>-->
-            <!--            <version>${spring.version}</version>  -->
-
             <!--		</dependency>-->
             <!--		<dependency>-->
             <!--			<groupId>org.liquibase</groupId>-->
