+++ kitchen-service/src/main/java/com/example/controllers/KithcenController.java
@@ -1,0 +1,16 @@
+package com.example.controllers;
+
+public class KithcenController {
+    public void createKithcen() {
+
+    }
+    public void changeKithcen() {
+
+    }
+    public void getKithcen() {
+
+    }
+    public void deleteKithcen() {
+
+    }
+}
