+++ kitchen-service/src/main/java/com/example/models/Dish.java
@@ -1,0 +1,4 @@
+package com.example.models;
+
+public class Dish {
+}
