+++ kitchen-service/src/main/java/com/example/models/Kitchen.java
@@ -1,0 +1,15 @@
+package com.example.models;
+
+import java.time.LocalDateTime;
+import java.util.List;
+
+public class Kitchen {
+    int id;
+    String nameKitchen;
+    String addressKitchen;
+    LocalDateTime timeWork;
+
+    int ratingKitchen;
+    Dish dish;
+    String statusOrder;
+}
