+++ kitchen-service/src/main/java/com/example/services/KitchenService.java
@@ -1,0 +1,16 @@
+package com.example.services;
+
+public class KitchenService {
+    public void createKithcen() {
+
+    }
+    public void changeKithcen() {
+
+    }
+    public void getKithcen() {
+
+    }
+    public void deleteKithcen() {
+
+    }
+}
