--- kitchen-service/src/main/java/ru/liga/KitchenServiceApplication.java
@@ -1,12 +1,0 @@
-package ru.liga;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class KitchenServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(KitchenServiceApplication.class,args);
-        System.out.println("hello from kithcen");
-    }
-}
