--- kitchen-service/src/main/java/ru/liga/controllers/KithcenController.java
@@ -1,16 +1,0 @@
-package ru.liga.controllers;
-
-public class KithcenController {
-    public void createKithcen() {
-
-    }
-    public void changeKithcen() {
-
-    }
-    public void getKithcen() {
-
-    }
-    public void deleteKithcen() {
-
-    }
-}
