--- kitchen-service/src/main/java/ru/liga/models/Dish.java
@@ -1,4 +1,0 @@
-package ru.liga.models;
-
-public class Dish {
-}
