--- kitchen-service/src/main/java/ru/liga/services/KitchenService.java
@@ -1,16 +1,0 @@
-package ru.liga.services;
-
-public class KitchenService {
-    public void createKithcen() {
-
-    }
-    public void changeKithcen() {
-
-    }
-    public void getKithcen() {
-
-    }
-    public void deleteKithcen() {
-
-    }
-}
