--- notification-service/src/main/java/ru/liga/NotificationServiceApplication.java
@@ -1,11 +1,0 @@
-package ru.liga;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class NotificationServiceApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(NotificationServiceApplication.class,args);
-    }
-}
