+++ order-service/src/main/java/com/example/StatusOrder.java
@@ -1,0 +1,7 @@
+package com.example;
+
+public enum StatusOrder {
+    ORDERISCREATED,
+    ORDERISCOOKING,
+    ORDERDELIVEREDCURRIER
+}
