+++ order-service/src/main/java/com/example/controllers/OrderController.java
@@ -1,0 +1,19 @@
+package com.example.controllers;
+
+public class OrderController {
+
+
+    public void createOrder() {
+
+    }
+    public void changeOrder() {
+
+    }
+    public void getOrder() {
+
+    }
+    public void deleteOrder() {
+
+    }
+
+}
