+++ order-service/src/main/java/com/example/models/Order.java
@@ -1,0 +1,10 @@
+package com.example.models;
+
+import java.time.LocalDateTime;
+
+public class Order {
+    int id;
+    String nameRestourant;
+    LocalDateTime timeOrder;
+    String items = "Тут будет список блюд в заказе";//сделать лист или сет
+}
