+++ order-service/src/main/java/com/example/services/OrderService.java
@@ -1,0 +1,16 @@
+package com.example.services;
+
+public class OrderService {
+    public void createOrder() {
+
+    }
+    public void changeOrder() {
+
+    }
+    public void getOrder() {
+
+    }
+    public void deleteOrder() {
+
+    }
+}
