--- order-service/src/main/java/ru/liga/StatusOrder.java
@@ -1,7 +1,0 @@
-package ru.liga;
-
-public enum StatusOrder {
-    ORDERISCREATED,
-    ORDERISCOOKING,
-    ORDERDELIVEREDCURRIER
-}
