--- order-service/src/main/java/ru/liga/controllers/OrderController.java
@@ -1,19 +1,0 @@
-package ru.liga.controllers;
-
-public class OrderController {
-
-
-    public void createOrder() {
-
-    }
-    public void changeOrder() {
-
-    }
-    public void getOrder() {
-
-    }
-    public void deleteOrder() {
-
-    }
-
-}
