--- order-service/src/main/java/ru/liga/models/Order.java
@@ -1,10 +1,0 @@
-package ru.liga.models;
-
-import java.time.LocalDateTime;
-
-public class Order {
-    int id;
-    String nameRestourant;
-    LocalDateTime timeOrder;
-    String items = "Тут будет список блюд в заказе";//сделать лист или сет
-}
