--- order-service/src/main/java/ru/liga/services/OrderService.java
@@ -1,16 +1,0 @@
-package ru.liga.services;
-
-public class OrderService {
-    public void createOrder() {
-
-    }
-    public void changeOrder() {
-
-    }
-    public void getOrder() {
-
-    }
-    public void deleteOrder() {
-
-    }
-}
