+++ src/main/java/com/example/ruligafoodservice/RuLigaFoodServiceApplication.java
@@ -1,0 +1,12 @@
+package com.example.ruligafoodservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RuLigaFoodServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(RuLigaFoodServiceApplication.class, args);
+    }
+
+}
