+++ src/test/java/com/example/ruligafoodservice/RuLigaFoodServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.ruligafoodservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RuLigaFoodServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
